Create hotel class and tests. Move methods from customer that make more sense for Hotel
