+++ src/customer.js
@@ -35,57 +35,6 @@ class Customer {
         return getCost
     }
 
-   getAvailableRoomByDate(date, allBookings, allRooms) {
-    let findRoomByDate = allBookings.bookings.reduce((acc, booking) => {
-        if(date !== booking.date) {
-            acc.push(booking.roomNumber)
-        }
-
-       return acc 
-    }, [])
-
-    // console.log("find", findRoomByDate)
-
-    const findRooms = allRooms.rooms.reduce((acc, room) => {
-        
-        findRoomByDate.forEach(number => {
-            if(number === room.number) {
-                acc.push(room)
-            }
-        })
-
-    return acc
-    }, [])
-    //  console.log("2", findRooms)
-    // console.log("availableRooms", availableRooms)
-
-        if(!findRooms) {
-            return 'We are terribly sorry, but there are no rooms available given your requests.'
-    }   else if (findRooms) {
-        this.availableRoomsByDate = findRooms
-        return this.availableRoomsByDate
-    }
-        // this.availableRoomsByDate = findRooms
-        // return this.availableRoomsByDate
-        ///NEED TO ADD SOME EDGE CASES HERE. WHAT IF THERE ARE NO ROOMS AVAILABLE BY DATE????
-
-   }
-
-
-   getAvailableRoomType(roomType) {
-    let availableRooms = this.availableRoomsByDate
-
-    const findByRoomType = availableRooms.filter(room => roomType === room.roomType)
-    console.log("result?", findByRoomType)
-
-    if(!findByRoomType) {
-        return 'We are terribly sorry, but there are no rooms available given your requests.'
-    } else if (findByRoomType) {
-        return findByRoomType
-    }
-
-   }
-
 
 }
